requirements.txt
xmltodict
beautifulsoup4
rfc3987
lark-parser
lxml
